refactor: rvr 64bit address ptr (#2928)
refactor address pointer related variables to 64bit instead of 32bit.
this PR modifies multiple files:
1) rvr core c and trace headers
2) ffi files
3) rvr code for extensions


reth benchmark result (original 32bit address space variables):
1) https://github.com/openvm-org/rvr-openvm/actions/runs/28121173062
2) https://github.com/openvm-org/rvr-openvm/actions/runs/28122726878

reth benchmark result (64bit address pointer variables):
1) https://github.com/openvm-org/rvr-openvm/actions/runs/28119229848
2) https://github.com/openvm-org/rvr-openvm/actions/runs/28122157089

By comparing "Exec Median (ms)" columns, it seems that 64bit version is
`10ms` faster or slower compared with the original 32bit version (in
both pure and metered modes). I am not sure why, it could be system
noises affecting it. Also, the comparison only compares execution speed,
excluding other possible important metrics when evaluating this pr.

resolves INT-8321

---------

+++ extensions/pairing/rvr/c/rvr_ext_pairing.h
@@ -7,10 +7,10 @@ struct RvState;
 
 /* HintFinalExp phantom: computes pairing final exponentiation hint and
  * sets the hint stream (implemented in Rust). */
-extern void rvr_ext_pairing_hint_final_exp_bn254(RvState* state, uint32_t rs1_val,
-                                                 uint32_t rs2_val);
+extern void rvr_ext_pairing_hint_final_exp_bn254(RvState* state, uint64_t rs1_val,
+                                                 uint64_t rs2_val);
 
-extern void rvr_ext_pairing_hint_final_exp_bls12_381(RvState* state, uint32_t rs1_val,
-                                                     uint32_t rs2_val);
+extern void rvr_ext_pairing_hint_final_exp_bls12_381(RvState* state, uint64_t rs1_val,
+                                                     uint64_t rs2_val);
 
 #endif /* RVR_EXT_PAIRING_H */
